% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamelift_operations.R
\name{gamelift_describe_vpc_peering_connections}
\alias{gamelift_describe_vpc_peering_connections}
\title{Retrieves information on VPC peering connections}
\usage{
gamelift_describe_vpc_peering_connections(FleetId)
}
\arguments{
\item{FleetId}{Unique identifier for a fleet.}
}
\description{
Retrieves information on VPC peering connections. Use this operation to
get peering information for all fleets or for one specific fleet ID.
}
\details{
To retrieve connection information, call this operation from the AWS
account that is used to manage the Amazon GameLift fleets. Specify a
fleet ID or leave the parameter empty to retrieve all connection
records. If successful, the retrieved information includes both active
and pending connections. Active connections identify the IpV4 CIDR block
that the VPC uses to connect.
\itemize{
\item CreateVpcPeeringAuthorization
\item DescribeVpcPeeringAuthorizations
\item DeleteVpcPeeringAuthorization
\item CreateVpcPeeringConnection
\item DescribeVpcPeeringConnections
\item DeleteVpcPeeringConnection
}
}
\section{Request syntax}{
\preformatted{svc$describe_vpc_peering_connections(
  FleetId = "string"
)
}
}

\keyword{internal}
